Especially for stateful models, introduce special api calling methods:
The reason is that while StatelesModel can be easily triggered by
RequestQueryResponse action on the server as it never triggers its
separate "side effect" (which calls ajax  etc.), in case of the
StatefulModel, we have no such control. It means that once the server
trigger the RequestQueryResponse action (so the models set their initial
values properly for the search page), Stateful models start to fetch
data via their api adapters which is not what we want. So currently,
we inject special methods for calling api into AppServices and on the
server, their immediately end any observable. But this cannot be forced,
so it is up to a developer not to forget this...